+++ src/mod.ts
@@ -1,2 +1,3 @@
 export * from "./utils.ts";
 export * from "./build.ts";
+export * from "./hash.ts";